Added the test bash script, the test flow and the loop file
This will probably be temporary but this is what I'm currently using for testing the mytimeout node. The test script is a bash script. It requires and mqtt server, the mosquitto_pub and mosquitto_sub scripts and the loop file gives an example command to run the script. The flow I'm using is in the flow.json file.